Puts search results all on the same page.
