+++ src/components/style.scss
@@ -26,3 +26,14 @@
 .footer-icons__icon {
   margin-left: 5px;
 }
+
+.geosuggest__item--active {
+  background: #267dc0;
+  color: #fff;
+}
+
+.geosuggest__suggests--hidden {
+  max-height: 0;
+  overflow: hidden;
+  border-width: 0;
+}
